# docs/devicemanager_swagger.yaml
basePath: /
definitions:
...
  models.Hub:
    properties:
      device_ids:
        items:
          type: string
        type: array
      device_local_ids:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Manager API
  version: "0.1"
paths:
...
  /helper/id:
    get:
      description: transforms short id to long id
      parameters:
      - description: short id
        in: query
        name: short_id
        required: true
        type: string
      - description: prefix added to generated long id
        in: query
        name: prefix
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
      security:
      - Bearer: []
      summary: transforms short id to long id
      tags:
      - helper
  /hubs:
    post:
      description: create hub
      parameters:
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Hub'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hub'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create hub
      tags:
      - create
      - hubs
  /hubs/{id}:
    delete:
      description: delete hub
      parameters:
      - description: Hub Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete hub
      tags:
      - delete
      - hubs
    get:
      description: get hub
      parameters:
      - description: Hub Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hub'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get hub
      tags:
      - get
      - hubs
    head:
      description: head hub
      parameters:
      - description: Hub Id
        in: path
        name: id
        required: true
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Hub'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hub'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: head hub
      tags:
      - head
      - hubs
    put:
      description: set hub
      parameters:
      - description: Hub Id
        in: path
        name: id
        required: true
        type: string
      - description: only admins may set user_id; overwrites hub.OwnerId; defaults
          to existing hub.OwnerId and falls back to user-id of requesting user if
          hub does not exist
        in: query
        name: user_id
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Hub'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hub'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set hub
      tags:
      - set
      - hubs
  /hubs/{id}/name:
    put:
      description: set hub name
      parameters:
      - description: Hub Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: name
        in: body
        name: message
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hub'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set hub name
      tags:
      - set
      - hubs
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"